Tackle issue 17 (enhancement) - make timesheet row sorting an automatically applied attribute of a timesheet. Editor views now have auto vs manual sorting. The old row movement or manual sorting options automatically disable auto sort.
JavaScript added to enable/disable buttons related to this - in turn that made
the move-up/move-down/remove row button section look strange as it didn't
enable/disable things in response to check box toggling, so added that too.

The change includes an appropriate migration which does not modify existing
data, because a nil/empty "auto_sort" field is used to mean "manual" - which
is of course the correct state for all timesheets predating this new feature.