The overdue full rewrite of the Report engine is complete. This version is very much faster than its predecessors (at least 5 times faster, often more). The "doc/app" documentation via RDoc describes the new API, as does in many ways the diff of the _tasks.html.erb, _comprehensive.html.erb and _users.html.erb partials.
Hand in hand with this, a rewrite of the Sections engine. The new code
is much cleaner and has a far easier to understand API. Again, see the
"doc/app" documentation and/or revision diffs for details. The diff of
the CSV report generator is particularly useful for seeing how it
changed for both the reports and sections API alterations.

Tweaked the CSS to improve general page appearance, especially line
wrapped headings.

General efficiency improvement for parts of the code that use the Task
model by eager loading associated projects and customers. Code that
interacts with a Task nearly always goes on to interact with its
Projects and Customers too.

Removed redundant CSV export code from ReportsController (it was put in
a report generator file in an earlier version but the old code in
ReportsController was overlooked).

Removed redundant "saved_reports_copy_controller.rb" file (the intended
functionality ended up part of SavedReportsController).

Fixed the "copy report" feature, which didn't actually copy reports (!).

Fixed the URLs for saved reports in the index view - the "user ID" part
of the URL isn't actually used for much right now, but ought to at least
be generated correctly.